backend/flask_app/requirements.txt
flask
flask_cors
sentinelhub
pillow
torch==1.10.1
torchvision==0.11.2
scipy